+++ java/com/turbovnc/rfb/CConnection.java
@@ -2,6 +2,7 @@
  * Copyright (C) 2011-2012 Brian P. Hinz
  * Copyright (C) 2012, 2014, 2016, 2018, 2020-2022 D. R. Commander.
  *                                                 All Rights Reserved.
+ * Copyright 2019 Pierre Ossman for Cendio AB
  *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -21,6 +22,8 @@
 
 package com.turbovnc.rfb;
 
+import java.nio.*;
+import java.nio.charset.*;
 import java.util.*;
 
 import com.turbovnc.network.*;
@@ -350,6 +353,25 @@ public abstract class CConnection extends CMsgHandler {
     clientSecTypeOrder = csto;
   }
 
+  // handleClipboardAnnounce() is called whenever the server's clipboard
+  // changes.  requestClipboard() should be called in order to access the
+  // actual data.
+
+  public abstract void handleClipboardAnnounce(boolean available);
+
+  // handleClipboardData() is called whenever the server has sent its clipboard
+  // data in response to a previous call to requestClipboard().  Note that this
+  // function might never be called if the clipboard data was no longer
+  // available when the server received the request.
+
+  public abstract void handleClipboardData(String data);
+
+  // handleClipboardRequest() is called whenever the server requests clipboard
+  // data from the client.  It will only be called after the client has first
+  // announced a clipboard change via announceClipboard().
+
+  public abstract void handleClipboardRequest();
+
   // Other methods
 
   public CMsgReader reader() { return reader; }
@@ -410,6 +432,176 @@ public abstract class CConnection extends CMsgHandler {
     throw new ErrorException("getUserPasswd() called in base class (this shouldn't happen.)");
   }
 
+  // announceClipboard() informs the server that the client's clipboard has
+  // changed.  The server may later request the clipboard data via
+  // handleClipboardRequest().
+
+  public void announceClipboard(boolean available)
+  {
+    if (!opts.sendClipboard)
+      return;
+
+    hasLocalClipboard = available;
+    unsolicitedClipboardAttempt = false;
+
+    // Attempt an unsolicited transfer?
+    if (available && cp.clipboardSize(RFB.EXTCLIP_FORMAT_UTF8) > 0 &&
+        (cp.clipboardFlags() & RFB.EXTCLIP_ACTION_PROVIDE) != 0) {
+      vlog.debug("Attempting unsolicited clipboard transfer...");
+      unsolicitedClipboardAttempt = true;
+      handleClipboardRequest();
+      return;
+    }
+
+    if ((cp.clipboardFlags() & RFB.EXTCLIP_ACTION_NOTIFY) != 0) {
+      writer().writeClipboardNotify(available ? RFB.EXTCLIP_FORMAT_UTF8 : 0);
+      return;
+    }
+
+    if (available)
+      handleClipboardRequest();
+  }
+
+  void handleClipboardCaps(int flags, int[] lengths)
+  {
+    int[] sizes = new int[]{ 0 };
+
+    super.handleClipboardCaps(flags, lengths);
+
+    writer().writeClipboardCaps(RFB.EXTCLIP_FORMAT_UTF8 |
+                                RFB.EXTCLIP_ACTION_REQUEST |
+                                RFB.EXTCLIP_ACTION_PEEK |
+                                RFB.EXTCLIP_ACTION_NOTIFY |
+                                RFB.EXTCLIP_ACTION_PROVIDE, sizes);
+  }
+
+  void handleClipboardNotify(int flags)
+  {
+    if (!opts.recvClipboard)
+      return;
+
+    serverClipboard = null;
+
+    if ((flags & RFB.EXTCLIP_FORMAT_UTF8) != 0) {
+      hasLocalClipboard = false;
+      handleClipboardAnnounce(true);
+    } else {
+      handleClipboardAnnounce(false);
+    }
+  }
+
+  void handleClipboardPeek(int flags)
+  {
+    if (!opts.sendClipboard)
+      return;
+
+    if ((cp.clipboardFlags() & RFB.EXTCLIP_ACTION_NOTIFY) != 0)
+      writer().writeClipboardNotify(hasLocalClipboard ?
+                                    RFB.EXTCLIP_FORMAT_UTF8 : 0);
+  }
+
+  void handleClipboardProvide(int flags, int[] lengths, byte[][] buffers)
+  {
+    if (!opts.recvClipboard)
+      return;
+
+    if ((flags & RFB.EXTCLIP_FORMAT_UTF8) == 0) {
+      vlog.debug("Ignoring Extended Clipboard provide message with " +
+                 "unsupported formats 0x" + Integer.toHexString(flags));
+      return;
+    }
+
+    if (buffers[0][lengths[0] - 1] == 0)
+      lengths[0]--;
+    serverClipboard =
+      Utils.convertLF(new String(buffers[0], 0, lengths[0]));
+
+    // FIXME: Should probably verify that this data was actually requested
+    handleClipboardData(serverClipboard);
+  }
+
+  void handleClipboardRequest(int flags)
+  {
+    if (!opts.sendClipboard)
+      return;
+
+    if ((flags & RFB.EXTCLIP_FORMAT_UTF8) == 0) {
+      vlog.debug("Ignoring Extended Clipboard request with unsupported " +
+                 "formats 0x" + Integer.toHexString(flags));
+      return;
+    }
+    if (!hasLocalClipboard) {
+      vlog.debug("Ignoring unexpected clipboard request");
+      return;
+    }
+    handleClipboardRequest();
+  }
+
+  // requestClipboard() requests clipboard data from the server.
+  // handleClipboardData() will be called once the data is available.
+
+  public void requestClipboard()
+  {
+    if (!opts.recvClipboard)
+      return;
+
+    if (serverClipboard != null) {
+      handleClipboardData(serverClipboard);
+      return;
+    }
+
+    if ((cp.clipboardFlags() & RFB.EXTCLIP_ACTION_REQUEST) != 0)
+      writer().writeClipboardRequest(RFB.EXTCLIP_FORMAT_UTF8);
+  }
+
+  // sendClipboardData() transfers the client's clipboard data to the server
+  // and should be called whenever the server has requested the clipboard data
+  // via handleClipboardRequest().
+
+  public void sendClipboardData(String data)
+  {
+    if (!opts.sendClipboard)
+      return;
+
+    if ((cp.clipboardFlags() & RFB.EXTCLIP_ACTION_PROVIDE) != 0) {
+      String filtered = Utils.convertCRLF(data);
+      int[] lengths = new int[1];
+      byte[][] datas = new byte[1][];
+
+      byte[] filteredBytes = filtered.getBytes();
+      lengths[0] = filteredBytes.length + 1;
+      datas[0] = new byte[filteredBytes.length + 1];
+      System.arraycopy(filteredBytes, 0, datas[0], 0, filteredBytes.length);
+
+      if (unsolicitedClipboardAttempt) {
+        unsolicitedClipboardAttempt = false;
+        if (lengths[0] > cp.clipboardSize(RFB.EXTCLIP_FORMAT_UTF8)) {
+          vlog.debug(lengths[0] +
+                     "-byte clipboard was too large for unsolicited transfer");
+          if ((cp.clipboardFlags() & RFB.EXTCLIP_ACTION_NOTIFY) != 0)
+            writer().writeClipboardNotify(RFB.EXTCLIP_FORMAT_UTF8);
+          return;
+        }
+      }
+
+      writer().writeClipboardProvide(RFB.EXTCLIP_FORMAT_UTF8, lengths, datas);
+    } else {
+      writer().writeClientCutText(data);
+    }
+  }
+
+  public void serverCutText(String str)
+  {
+    if (!opts.recvClipboard)
+      return;
+
+    hasLocalClipboard = false;
+
+    serverClipboard = str;
+
+    handleClipboardAnnounce(true);
+  }
+
   InStream is;
   OutStream os;
   protected CMsgReader reader;
@@ -425,6 +617,10 @@ public abstract class CConnection extends CMsgHandler {
   protected Socket sock;
   boolean alreadyPrintedVersion, alreadyPrintedSecurity;
   boolean alreadyPrintedSecurityResult;
+
+  private boolean hasLocalClipboard, unsolicitedClipboardAttempt;
+  private String serverClipboard;
+
   // CHECKSTYLE VisibilityModifier:OFF
   public Options opts;
   // CHECKSTYLE VisibilityModifier:ON