+++ java/com/turbovnc/rfb/ConnParams.java
@@ -1,6 +1,8 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
- * Copyright (C) 2011-2012, 2015, 2018 D. R. Commander.  All Rights Reserved.
+ * Copyright (C) 2011-2012, 2015, 2018, 2022 D. R. Commander.
+ *                                           All Rights Reserved.
  * Copyright (C) 2012 Brian P. Hinz
+ * Copyright 2019 Pierre Ossman for Cendio AB
  *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -110,6 +112,33 @@ public class ConnParams {
 
   boolean done;
 
+  public int clipboardFlags() { return clipFlags; }
+
+  public int clipboardSize(int format)
+  {
+    for (int i = 0; i < 16; i++) {
+      if ((1 << i) == format)
+        return clipSizes[i];
+    }
+
+    throw new ErrorException("Invalid clipboard format 0x" +
+                             Integer.toHexString(format));
+  }
+
+  public void setClipboardCaps(int flags, int[] lengths)
+  {
+    int i, num;
+
+    clipFlags = flags;
+
+    num = 0;
+    for (i = 0; i < 16; i++) {
+      if ((flags & (1 << i)) == 0)
+        continue;
+      clipSizes[i] = lengths[num++];
+    }
+  }
+
   // CHECKSTYLE VisibilityModifier:OFF
   public boolean supportsDesktopResize;
   public boolean supportsExtendedDesktopSize;
@@ -129,4 +158,6 @@ public class ConnParams {
   private int[] encodings;
   private StringBuilder verStr;
   private int verStrPos;
+  private int clipFlags;
+  private int[] clipSizes = new int[16];
 }